Add admin layout in GoHTML
New HTML layout introduced for admin section with related JavaScript and CSS assets. The layout includes necessary templates, a minimalistic navbar and a title to be filled in. It also includes the admin.bundle.min.js, which bundles all relevant JavaScript for the admin functionality, potentially speeding up loading times.