Remove unnecessary frontend live/public js/admin/app.js file
This commit removes the frontend live/public js/admin/app.js file which is no longer needed. This is a part of a cleanup process to remove unused or unnecessary files to make the codebase more maintainable and cleaner. Its deletion does not impact any working functionalities of the project.